+++ migration/index.spec.ts
@@ -18,7 +18,14 @@ describe("Migration", () => {
   });
 
   tester.run("v26", migration.rules.v26, {
-    valid: [`const routing = { "get /": someEndpoint };`],
+    valid: [
+      `const routing = { "get /": someEndpoint };`,
+      `factory.build({ handler: async ({ ctx }) => {} });`,
+      `factory.addContext();`,
+      `new Middleware({ handler: async ({ ctx }) => {} });`,
+      `new ResultHandler({ handler: ({ ctx }) => {} });`,
+      `testMiddleware({ ctx: {} });`,
+    ],
     invalid: [
       {
         name: "basic DependsOnMethod",
@@ -92,6 +99,61 @@ describe("Migration", () => {
               from: "new DependsOnMethod(...)",
               to: "its argument object and append its keys with ' /'",
             },
+          },
+        ],
+      },
+      {
+        name: "options in handler",
+        code: `factory.build({ handler: async ({ options }) => {} });`,
+        output: `factory.build({ handler: async ({ ctx }) => {} });`,
+        errors: [
+          {
+            messageId: "change",
+            data: { subject: "property", from: "options", to: "ctx" },
+          },
+        ],
+      },
+      {
+        name: "renamed options in handler",
+        code: `new Middleware({ handler: async ({ options: ttt }) => {} });`,
+        output: `new Middleware({ handler: async ({ ctx: ttt }) => {} });`,
+        errors: [
+          {
+            messageId: "change",
+            data: { subject: "property", from: "options", to: "ctx" },
+          },
+        ],
+      },
+      {
+        name: "destructed options in handler",
+        code: `new ResultHandler({ handler: ({ options: { method } }) => {} });`,
+        output: `new ResultHandler({ handler: ({ ctx: { method } }) => {} });`,
+        errors: [
+          {
+            messageId: "change",
+            data: { subject: "property", from: "options", to: "ctx" },
+          },
+        ],
+      },
+      {
+        name: "addOptions method",
+        code: `factory.addOptions(() => {});`,
+        output: `factory.addContext(() => {});`,
+        errors: [
+          {
+            messageId: "change",
+            data: { subject: "method", from: "addOptions", to: "addContext" },
+          },
+        ],
+      },
+      {
+        name: "testMiddleware options property",
+        code: `testMiddleware({ options: {} });`,
+        output: `testMiddleware({ ctx: {} });`,
+        errors: [
+          {
+            messageId: "change",
+            data: { subject: "property", from: "options", to: "ctx" },
           },
         ],
       },